Moved function to utils.
